Fix CVE-2019-0201
Why should update to 3.5.5 due to: CVE-2019-0201

More information
moderate severity
Vulnerable versions: >= 3.5.0, < 3.5.5
Patched version: 3.5.5

An issue is present in Apache ZooKeeper 1.0.0 to 3.4.13 and 3.5.0-alpha to 3.5.4-beta. ZooKeeper?s getACL() command doesn?t check any permission when retrieves the ACLs of the requested node and returns all information contained in the ACL Id field as plaintext string. DigestAuthenticationProvider overloads the Id field with the hash value that is used for user authentication. As a consequence, if Digest Authentication is in use, the unsalted hash value will be disclosed by getACL() request for unauthenticated or unprivileged users

source : https://github.com/apache/bookkeeper/issues/2106